// csm.web.app/src/app/project/project-common/enums/task-entity-permission.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2023
 *
 * *************************************************************************
 */

export enum TaskEntityPermissionEnum {
    Assign = 'ASSIGN',
    Unassign = 'UNASSIGN',
    Send = 'SEND',
    Delete = 'DELETE',
    Reset = 'RESET',
    Start = 'START',
    Close = 'CLOSE',
    Accept = 'ACCEPT',
    Topic = 'TOPIC',
    Update = 'UPDATE',
    UpdateConstraints = 'UPDATE_CONSTRAINTS',
}